sensor-data-scala cloudflow example from docs, expanded into sbt multi-module project.
