+++ .scalafmt.conf
@@ -1,0 +1,44 @@
+version = "2.3.2"
+
+style = defaultWithAlign
+
+align                               = more
+align.openParenCallSite             = true
+align.openParenDefnSite             = true
+rewrite.rules                       = [AvoidInfix,RedundantBraces,RedundantParens,SortImports]
+danglingParentheses                 = true
+docstrings                          = JavaDoc
+maxColumn                           = 140
+unindentTopLevelOperators           = true
+rewrite.neverInfix.excludeFilters   = [
+  and
+  min
+  max
+  until
+  to
+  by
+  eq
+  ne
+  "should.*"
+  "contain.*"
+  "must.*"
+  in
+  ignore
+  be
+  taggedAs
+  thrownBy
+  synchronized
+  have
+  when
+  size
+  only
+  noneOf
+  oneElementOf
+  noElementsOf
+  atLeastOneElementOf
+  atMostOneElementOf
+  allElementsOf
+  inOrderElementsOf
+  theSameElementsAs
+]
+spaces.inImportCurlyBraces          = true
