--- app/src/main/blueprint/blueprint.conf
@@ -1,16 +1,0 @@
-blueprint {
-  streamlets {
-    http-ingress = sensordata.SensorDataHttpIngress
-    metrics = sensordata.SensorDataToMetrics
-    validation = sensordata.MetricsValidation
-    valid-logger = sensordata.ValidMetricLogger
-    invalid-logger = sensordata.InvalidMetricLogger
-  }
-
-  connections {
-    http-ingress.out = [metrics.in]
-    metrics.out = [validation.in]
-    validation.invalid = [invalid-logger.in]
-    validation.valid = [valid-logger.in]
-  }
-}
