+++ sensor-data-app/src/main/resources/local.conf
@@ -1,0 +1,6 @@
+file-ingress {
+  source-data-mount="/tmp/cloudflow"
+}
+valid-logger {
+  log-level = "debug"
+}
