+++ sensor-data-app/src/main/resources/logback.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+    <logger name="akka" level="WARN" />
+    <logger name="org.apache" level="WARN" />
+    <logger name="kafka" level="WARN" />
+    <logger name="org.spark_project.jetty" level="WARN" />
+
+    <root level="INFO">
+        <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+            <target>System.out</target>
+            <encoder>
+                <pattern>%d{ISO8601} %-5level [%logger{0}] - %msg%n</pattern>
+            </encoder>
+        </appender>
+    </root>
+</configuration>
