+++ sensor-data-ingress/src/main/scala/package.scala
@@ -1,0 +1,5 @@
+import java.time.Instant
+
+package object sensordata {
+  implicit def toInstant(millis: Long): Instant = Instant.ofEpochMilli(millis)
+}
