+++ sensor-data-logging/src/main/scala/sensordata/InvalidMetricLogger.scala
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2016-2020 Lightbend Inc. <https://www.lightbend.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sensordata
+
+import cloudflow.akkastream.AkkaStreamlet
+import cloudflow.akkastream.scaladsl.{FlowWithCommittableContext, RunnableGraphStreamletLogic}
+import cloudflow.streamlets.StreamletShape
+import cloudflow.streamlets.avro.AvroInlet
+
+class InvalidMetricLogger extends AkkaStreamlet {
+  val inlet: AvroInlet[InvalidMetric] = AvroInlet[InvalidMetric]("in")
+  val shape: StreamletShape = StreamletShape.withInlets(inlet)
+
+  override def createLogic = new RunnableGraphStreamletLogic() {
+    private val flow = FlowWithCommittableContext[InvalidMetric].map { invalidMetric =>
+      system.log.warning(s"Invalid metric detected! $invalidMetric")
+      invalidMetric
+    }
+
+    def runnableGraph =
+      sourceWithOffsetContext(inlet).via(flow).to(committableSink)
+  }
+}
