+++ sensor-data-model/data/10-storm.json
@@ -1,0 +1,9 @@
+{
+    "deviceId": "c75cb448-df0e-4692-8e06-0321b7703992",
+    "timestamp": 1495545346279,
+    "measurements": {
+        "power": 1.7,
+        "rotorSpeed": 23.4,
+        "windSpeed": 100.1
+    }
+}
