+++ sensor-data-model/data/12-hurricane.json
@@ -1,0 +1,9 @@
+{
+    "deviceId": "c75cb448-df0e-4692-8e06-0321b7703992",
+    "timestamp": 1495546546279,
+    "measurements": {
+        "power": 1.7,
+        "rotorSpeed": 78.3,
+        "windSpeed": 129.4
+    }
+}
