+++ sensor-data-model/data/future-data.json
@@ -1,0 +1,41 @@
+[
+  {
+      "deviceId": "c75cb448-df0e-4692-8e06-0321b7703992",
+      "timestamp": 1495545346279,
+      "measurements": {
+          "power": 1.7,
+          "rotorSpeed": 3.9,
+          "windSpeed": 25.3
+      }
+  },
+  {
+    "deviceId": "c75cb448-df0e-4692-8e06-0321b7703992",
+    "timestamp": 1495545346279,
+    "measurements": {
+        "power": -1.7,
+        "rotorSpeed": 3.9,
+        "windSpeed": 25.3
+    }
+  }
+  ,
+  {
+    "deviceId": "c75cb448-df0e-4692-8e06-0321b7703992",
+    "timestamp": 3134584800000,
+    "measurements": {
+        "power": 1.7,
+        "rotorSpeed": 3.9,
+        "windSpeed": 100.1
+    }
+  },
+    {
+    "deviceId": "c75cb448-df0e-4692-8e06-0321b7703992",
+    "timestamp": 3134584800000,
+    "measurements": {
+        "power": 1.7,
+        "rotorSpeed": 3.9,
+        "windSpeed": 100.1
+    }
+  }
+]
+
+
