+++ sensor-data-model/src/main/avro/InvalidMetric.avsc
@@ -1,0 +1,13 @@
+{
+    "namespace": "sensordata",
+    "type": "record",
+    "name": "InvalidMetric",
+    "fields":[
+         {
+            "name": "metric", "type": "sensordata.Metric"
+         },
+         {
+            "name": "error", "type": "string"
+         }
+    ]
+}
