+++ sensor-data-model/src/main/avro/Measurements.avsc
@@ -1,0 +1,16 @@
+{
+    "namespace": "sensordata",
+    "type": "record",
+    "name": "Measurements",
+    "fields":[
+         {
+            "name": "power", "type": "double"
+         },
+         {
+            "name": "rotorSpeed", "type": "double"
+         },
+         {
+            "name": "windSpeed", "type": "double"
+         }
+    ]
+}
