+++ sensor-data-model/src/main/avro/SensorData.avsc
@@ -1,0 +1,24 @@
+{
+    "namespace": "sensordata",
+    "type": "record",
+    "name": "SensorData",
+    "fields":[
+         {
+            "name": "deviceId",
+            "type": {
+                "type": "string",
+                "logicalType": "uuid"
+            }
+         },
+         {
+            "name": "timestamp",
+            "type": {
+                "type": "long",
+                "logicalType": "timestamp-millis"
+            }
+         },
+         {
+            "name": "measurements", "type": "sensordata.Measurements"
+         }
+    ]
+}
