+++ sensor-data-validation/src/main/scala/sensordata/SensorDataUtils.scala
@@ -1,0 +1,5 @@
+package sensordata
+
+object SensorDataUtils {
+  def isValidMetric(m: Metric) = m.value >= 0.0
+}
