--- sensor-data/src/main/avro/InvalidMetric.avsc
@@ -1,13 +1,0 @@
-{
-    "namespace": "sensordata",
-    "type": "record",
-    "name": "InvalidMetric",
-    "fields":[
-         {
-            "name": "metric", "type": "sensordata.Metric"
-         },
-         {
-            "name": "error", "type": "string"
-         }
-    ]
-}
