--- sensor-data/src/main/avro/Measurements.avsc
@@ -1,16 +1,0 @@
-{
-    "namespace": "sensordata",
-    "type": "record",
-    "name": "Measurements",
-    "fields":[
-         {
-            "name": "power", "type": "double"
-         },
-         {
-            "name": "rotorSpeed", "type": "double"
-         },
-         {
-            "name": "windSpeed", "type": "double"
-         }
-    ]
-}
