--- sensor-data/src/main/avro/SensorData.avsc
@@ -1,24 +1,0 @@
-{
-    "namespace": "sensordata",
-    "type": "record",
-    "name": "SensorData",
-    "fields":[
-         {
-            "name": "deviceId",
-            "type": {
-                "type": "string",
-                "logicalType": "uuid"
-            }
-         },
-         {
-            "name": "timestamp",
-            "type": {
-                "type": "long",
-                "logicalType": "timestamp-millis"
-            }
-         },
-         {
-            "name": "measurements", "type": "sensordata.Measurements"
-         }
-    ]
-}
