--- sensor-data/src/main/scala/sensordata/InvalidMetricLogger.scala
@@ -1,37 +1,0 @@
-/*
- * Copyright (C) 2016-2020 Lightbend Inc. <https://www.lightbend.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sensordata
-
-import akka.stream.scaladsl.RunnableGraph
-import cloudflow.akkastream._
-import cloudflow.akkastream.scaladsl._
-import cloudflow.streamlets._
-import cloudflow.streamlets.avro._
-
-class InvalidMetricLogger extends AkkaStreamlet {
-  val inlet: AvroInlet[InvalidMetric] = AvroInlet[InvalidMetric]("in")
-  val shape: StreamletShape = StreamletShape.withInlets(inlet)
-
-  override def createLogic: AkkaStreamletLogic = new RunnableGraphStreamletLogic() {
-    private val flow = FlowWithCommittableContext[InvalidMetric].map { invalidMetric =>
-      system.log.warning(s"Invalid metric detected! $invalidMetric")
-      invalidMetric
-    }
-
-    def runnableGraph: RunnableGraph[_] = sourceWithOffsetContext(inlet).via(flow).to(committableSink)
-  }
-}
