--- sensor-data/src/main/scala/sensordata/SensorDataHttpIngress.scala
@@ -1,31 +1,0 @@
-/*
- * Copyright (C) 2016-2020 Lightbend Inc. <https://www.lightbend.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sensordata
-
-import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
-import cloudflow.akkastream._
-import cloudflow.akkastream.util.scaladsl._
-
-import cloudflow.streamlets._
-import cloudflow.streamlets.avro._
-import json.SensorDataJsonSupport._
-
-class SensorDataHttpIngress extends AkkaServerStreamlet {
-  val out: AvroOutlet[SensorData] = AvroOutlet[SensorData]("out").withPartitioner(RoundRobinPartitioner)
-  def shape: StreamletShape = StreamletShape.withOutlets(out)
-  override def createLogic: AkkaStreamletLogic = HttpServerLogic.default(this, out)
-}
