--- sensor-data/src/main/scala/sensordata/json/MeasurementsJsonSupport.scala
@@ -1,8 +1,0 @@
-package sensordata.json
-
-import sensordata.Measurements
-import spray.json.{DefaultJsonProtocol, RootJsonFormat}
-
-object MeasurementsJsonSupport extends DefaultJsonProtocol {
-  implicit val measurementFormat: RootJsonFormat[Measurements] = jsonFormat3(Measurements.apply)
-}
